+++ src/oaklib/implementations/__init__.py
@@ -36,7 +36,7 @@ from oaklib.implementations.ontoportal.ontoportal_implementation_base import (
     OntoPortalImplementationBase,
 )
 from oaklib.implementations.pronto.pronto_implementation import ProntoImplementation
-from oaklib.implementations.rustsim.rustsim_implementation import RustSimImplementation
+from oaklib.implementations.semsimian.semsimian_implementation import RustSimImplementation
 from oaklib.implementations.simpleobo.simple_obo_implementation import (
     SimpleOboImplementation,
 )
@@ -130,7 +130,7 @@ def get_implementation_resolver() -> ClassResolver[OntologyInterface]:
             "oak": OakMetaModelImplementation,
             "cx": CXImplementation,
             "ndexbio": CXImplementation,
-            "rustsim": RustSimImplementation,
+            "semsimian": RustSimImplementation,
         }
     )
