+++ tox.ini
@@ -24,7 +24,7 @@ commands =
 deps =
     pytest
     coverage
-    rustsim
+    semsimian
 
 [testenv:lint]
 deps =
